Protect the endpoints with jwt token

+++ backend/controllers/movieController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Movie = require('../models/movieModel')
+const User = require('../models/userModel')
 
 // @desc Get movies
 // @route GET /api/movies
 // @access Privet
 const getMovies = asyncHandler(async (req, res) => {
-  const movies = await Movie.find()
+  const movies = await Movie.find({ user: req.user.id })
   res.status(200).json(movies)
 })
 
@@ -21,6 +22,7 @@ const setMovie = asyncHandler(async (req, res) => {
     const movie = await Movie.create({
       title: req.body.title,
       year: req.body.year,
+      user: req.user.id,
     })
     res.status(200).json({ movie })
   }
@@ -35,6 +37,15 @@ const updateMovie = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Movie not found')
   } else {
+    const user = await User.findById(req.user.id)
+    if (!user) {
+      res.status(400)
+      throw new Error('User not found')
+    }
+    if (movie.user.toString() !== user.id) {
+      res.status(401)
+      throw new Error('User not authorized')
+    }
     const updatedMovie = await Movie.findByIdAndUpdate(
       req.params.id,
       req.body,
@@ -55,6 +66,15 @@ const deleteMovie = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Movie not found')
   } else {
+    const user = await User.findById(req.user.id)
+    if (!user) {
+      res.status(400)
+      throw new Error('User not found')
+    }
+    if (movie.user.toString() !== user.id) {
+      res.status(401)
+      throw new Error('User not authorized')
+    }
     const deletedMovie = await Movie.findByIdAndDelete(req.params.id)
     res.status(200).json({ message: `Movie ${deletedMovie.title} deleted` })
   }
